refactor(auth): split agent action input mapping
- move delegated capability input decoders into an agent-local helper
- leave auth options focused on runner construction and Better Auth wiring
- reduce options.ts hotspot score from 25.1 to 17.4 and density from 0.13 to 0.09
